Update Hyperelliptic2.magma
Fixed a serious error in the function fourpoints31. The earlier version gave incorrect results when run on certain fields K containing cube roots of 1. The error does not affect any of the computations involved in enumerating hyperelliptic curves of genus 2, because fourpoints31 is not used in any of these computations. It *does* affect the computations that enumerate
hyperelliptic curves of genus 3 using the programs in Hyperelliptic3.magma.